bmc ipcfg: initial commit
Allows configuring the ip info in u-boot's bootargs.
Updates both QSPI0 and QSPI1 environments.
Stores persistent ip information in a persistent block
in each QSPI, so that the IP info will survive image upgrades
and changes to the environment.

Work remaining is to calculate the env CRC, actuall do the
write to the env blocks, and to call the reconcile function
on boot-up and as the final step of an image upgrade.

Signed-off-by: Mark Pleso <mpleso@platinasystems.com>