ubisetup: Initial commit
Initial commit of UBI support. In order to facilitate migration
from the old static partitioning scheme to the new scheme, upon
booting we check to see whether UBI has been set up or not.

If UBI is not set up, we will read the static partitions into memory
and reformat the space for UBI. We then write back the data from
the static partitions into files in the UBIFS.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>