bmc: Fix ITB support, add size checks
Set the u-boot environment variable fdt_high and initrd_high because
by default u-boot places the fdt and initrd extracted from ITB files
at the top of memory. The kernel (on ARM) can only access these
items if they are in the first 512MiB of physical memory, so this fails
on systems with > 512MiB of RAM.

Check the size of the images we are creating and error if they exceed
the size. The upgrade code does not check and oversized images can
brick a device.

Add U-boot environment support for booting in ITB mode, and if that
fails, falling back to legacy mode.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>