goes: Enable sshd with FailSafe disabled
To allow for more global testing, and to deal with the size growth
from pulling in ssh and the associated crypto libraries, add ssh
to the BMC image.

Since there is not persistent storage, do not enable FailSafe. FailSafe
is the feature that leaves the box open when there are no configured
authorized keys.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>