upgrade: Changes for new flash layout
The upgrade command now operates on the currently selected QSPI device.
In order to change selected device (to upgrade QSPI1, for example), the
qspi command is used.

If upgrading to a modern version, make changes to the UBIFS filesystem.

If downgrading to a legacy version, write the raw partitions as is
expected by earlier firmware.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>