upgrade: ensure that UBI-based versions have ubi mounted
If the selected QSPI is UBI-formatted, require that the UBI partition
be mounted for upgrades. In the case that we are downgrading to a pre-UBI
legacy version, use the -legacy option to allow for a downgrade.

Undocumented feature: specifying -legacy with a modern version causes
the modern version to be installed as if we are upgrading. This can
be useful for testing or if the ubi partition is corrupted.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>