bmc upgrade: Ensure partitions are mounted as they should be
The 'reboot' command (or umount -a, which it calls) can fail while
unmounting one of the UBI partitions (because a user did a 'cd' to it
and issued the reboot or umount).

The 'upgrade' command only checks that the UBI volume is mounted somewhere;
it doesn't make sure that the required bind mounts are all still present.

Add a check to make sure that all of the persistent partitions are
mounted prior to an upgrade if this is UBI-formatted.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>