ucd9090d: Fix Vout reply offsets
Previously, each command added two requests to the I2C request queue
(one to set the mux, and one for the actual command). When I recalculated
the offsets, I didn't calculate properly for each command, instead assuming
that there was only one set of the mux at the beginning of the command.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>